requirements.txt
faiss-cpu
langchain~=0.3.27
langchain-community~=0.3.30
langchain-groq
langchain-huggingface
pypdf
sentence-transformers
streamlit~=1.50.0

pip~=25.2
attrs~=25.3.0
distro~=1.9.0
protobuf~=6.32.1
pillow~=11.3.0
filelock~=3.19.1
python-dotenv~=1.1.1